@@ -1,14 +1,14 @@
-// Mocks generated by Mockito 5.3.0 from annotations
-// in webview_flutter/test/v4/webview_cookie_manager_test.dart.
+// Mocks generated by Mockito 5.3.2 from annotations
+// in webview_flutter/test/webview_cookie_manager_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i4;
 
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:webview_flutter_platform_interface/v4/src/platform_webview_cookie_manager.dart'
+import 'package:webview_flutter_platform_interface/src/platform_webview_cookie_manager.dart'
     as _i3;
-import 'package:webview_flutter_platform_interface/v4/src/webview_platform.dart'
+import 'package:webview_flutter_platform_interface/src/webview_platform.dart'
     as _i2;
 
 // ignore_for_file: type=lint
@@ -25,8 +25,12 @@ import 'package:webview_flutter_platform_interface/v4/src/webview_platform.dart'
 class _FakePlatformWebViewCookieManagerCreationParams_0 extends _i1.SmartFake
     implements _i2.PlatformWebViewCookieManagerCreationParams {
   _FakePlatformWebViewCookieManagerCreationParams_0(
-      Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
 }
 
 /// A class which mocks [PlatformWebViewCookieManager].
@@ -40,17 +44,28 @@ class MockPlatformWebViewCookieManager extends _i1.Mock
 
   @override
   _i2.PlatformWebViewCookieManagerCreationParams get params =>
-      (super.noSuchMethod(Invocation.getter(#params),
-              returnValue: _FakePlatformWebViewCookieManagerCreationParams_0(
-                  this, Invocation.getter(#params)))
-          as _i2.PlatformWebViewCookieManagerCreationParams);
+      (super.noSuchMethod(
+        Invocation.getter(#params),
+        returnValue: _FakePlatformWebViewCookieManagerCreationParams_0(
+          this,
+          Invocation.getter(#params),
+        ),
+      ) as _i2.PlatformWebViewCookieManagerCreationParams);
   @override
-  _i4.Future<bool> clearCookies() =>
-      (super.noSuchMethod(Invocation.method(#clearCookies, []),
-          returnValue: _i4.Future<bool>.value(false)) as _i4.Future<bool>);
+  _i4.Future<bool> clearCookies() => (super.noSuchMethod(
+        Invocation.method(
+          #clearCookies,
+          [],
+        ),
+        returnValue: _i4.Future<bool>.value(false),
+      ) as _i4.Future<bool>);
   @override
   _i4.Future<void> setCookie(_i2.WebViewCookie? cookie) => (super.noSuchMethod(
-      Invocation.method(#setCookie, [cookie]),
-      returnValue: _i4.Future<void>.value(),
-      returnValueForMissingStub: _i4.Future<void>.value()) as _i4.Future<void>);
+        Invocation.method(
+          #setCookie,
+          [cookie],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
 }
